Extract the last three pieces of PuzzleComponents
