+++ imports/client/components/puzzleShape.js
@@ -1,0 +1,3 @@
+const puzzleShape = Schemas.Puzzles.asReactPropTypes();
+
+export default puzzleShape;
