+++ imports/client/components/tagShape.js
@@ -1,0 +1,4 @@
+const tagShape = Schemas.Tags.asReactPropTypes();
+
+export default tagShape;
+
